Refactor insert and delete element functions
Modified the Insert and delete functions. Some of the stuff that occurred in both was moved to a separate function, since the share similarities, which can be done in that new function. Also added in detection for numbers outside of the array size to that function.

Also moved the which_array_to_modify function to inside the delete and insert while loops.

And added some test case ideas to the bottom.